refactor(rational): split rational into core / manipulators / iostream (#1334)
Phase 2 of the header layering, continuing after edecimal (#1423).

    core.hpp          arithmetic, no streams          59,500 lines, 0 I/O-family
    manipulators.hpp  type_tag / to_binary / to_native
    iostream.hpp      operator<< / operator>>
    rational.hpp      everything, as before           77,902 lines

The umbrella was 77,828 lines pulling all five stream headers. The core is 24% smaller
and pulls none -- the largest reduction of the rollout so far, because two headers
underneath it were split first: native/manipulators.hpp (for scale()) and blockdigit.hpp.

rational_impl.hpp had included internal/blockbinary/manipulators.hpp, which pulls four of
the five. Its only consumer was to_binary(rational<>), a manipulator, so that include
moved to manipulators.hpp along with the function. The impl now takes ieee754_core.hpp
and manipulators_core.hpp instead of the full ieee754.hpp: it wants extractFields,
ieee754_parameter, scale and find_msb, all of which are bit manipulation.

Five std::cerr diagnostics -- one divide-by-zero and two each of overflow and underflow --
became fprintf(stderr, ...). The overflow/underflow pair interpolated an int, so they
carry %d; the friend declarations need only <iosfwd>. <sstream> and <string> came out of
the impl entirely: nothing there uses them once the text layer has moved.

rational/manipulators.hpp had NO includes while using std::string, std::stringstream,
std::is_same_v and the base-format manipulators -- the same defect class as #1422. It now
states its own dependencies, including native/integers.hpp for the to_binary(Integer,bool)
fallback that to_binary(blockdigit) reaches on the base-2 path.

arithmetic_traits.hpp and number_traits_reports.hpp stay at the umbrella, not the core:
they build report strings and pull <sstream>/<iomanip> by design.

Verified: every api.cpp in the tree, 55 pass 0 fail, on gcc and clang; the full
static/rational, elastic/rational, internal/blockdigit and static/positional suites build
and run clean; hardened #1389 sweep clean across the subtree; ASCII clean.
