+++ include/sw/universal/number/rational/attributes.hpp
@@ -5,6 +5,11 @@
 // SPDX-License-Identifier: MIT
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
+#include <string>       // std::string
+#include <sstream>     // std::stringstream
+#include <iomanip>     // std::setw
+#include <type_traits> // std::enable_if_t
+#include <universal/number/rational/core.hpp>
 #include <cmath> // for std:pow()
 
 namespace sw { namespace universal {
