+++ include/sw/universal/number/rational/rational_fwd.hpp
@@ -5,6 +5,7 @@
 // SPDX-License-Identifier: MIT
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
+#include <cstdint>      // uint8_t, the default block type
 
 namespace sw { namespace universal {
